+++ md5.h
@@ -92,7 +92,7 @@ public:
 	/* Construct a MD5 object with a string. */
 	MD5(const string& message);
 
-	MD5(const char* message, int len);
+	MD5(const char* message, size_t len);
 
 	/* Generate md5 digest. */
 	const byte* getDigest();
